fix: align speckv calibration dtype with bfloat16 runs

+++ R-KV/weian_development/rkv_sparse_round_calibrate.py
@@ -91,7 +91,7 @@ def parse_args() -> argparse.Namespace:
     parser.add_argument(
         "--dtype",
         type=str,
-        default="float16",
+        default="bfloat16",
         choices=["float16", "bfloat16", "float32"],
         help="Computation dtype for forward capture.",
     )
